feat: add new habits
Implemented creation of new habits for a theme. Adding habits occurs by
clicking the floating action button in the bottom right, which displays
the Add Habit/Theme dialog. Adding a theme is not available at this
time.

Once on the dialog, the following steps are necessary for a new habit:

 1. Select theme to add to
 2. Provide title for new habit
 3. Click Add

The dialog closes and the habit appears in the user interface.